refactor EnumerationItem cruds
